% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-mcp.R
\name{execute_mcp_tool}
\alias{execute_mcp_tool}
\title{Execute an MCP tool}
\usage{
execute_mcp_tool(tool_def, arguments = list())
}
\arguments{
\item{tool_def}{List. The tool definition (e.g. one of the items returned by \code{mcp_tools()}).}

\item{arguments}{List. The arguments to pass to the tool}
}
\value{
The result of the tool execution
}
\description{
This is meant to test MCP tools manually, simulating a tool call from an LLM.
}
