% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{flat_list}
\alias{flat_list}
\title{Flatten a list of elements to a single-depth list}
\usage{
flat_list(...)
}
\arguments{
\item{...}{One or more elements or lists to flatten}
}
\value{
List with all elements at depth 1
}
\description{
Flatten a list of elements to a single-depth list
}
