% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-mcp.R
\name{get_mcp_resource}
\alias{get_mcp_resource}
\title{Get an MCP resource from a list of resource definitions (convenience function)}
\usage{
get_mcp_resource(resource_defs, resource_uri)
}
\arguments{
\item{resource_defs}{List. A list of resource definitions}

\item{resource_uri}{Character. The URI of the resource to get}
}
\value{
The resource definition
}
\description{
Get an MCP resource from a list of resource definitions (convenience function)
}
