% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-mcp.R
\name{read_mcp_resource}
\alias{read_mcp_resource}
\title{Read an MCP resource}
\usage{
read_mcp_resource(resource_def)
}
\arguments{
\item{resource_def}{List. The resource definition (e.g. one of the items returned by \code{mcp_resources()}).}
}
\value{
The content of the resource
}
\description{
Read an MCP resource
}
